Fix .omr replay with auto-run type command
When the auto-run feature is enabled, inserting a .cas file will
schedule some type commands. These commands end up in the .omr replay
file. Though when replaying those command it seems like the ENTER key is
no longer part of the replay.

These type commands end with a '\r' character (the enter key, to
actually execute the typed stuff on the BASIC prompt). This character
does get included in the .omr file (*1), but after re-loading it
disappeared. The problem was that we parsed .omr files (which are
(compressed) XML files) with the 'trimWhitespace' option.

Removing the 'trimWhitespace' option fixes the problem. I don't believe
we ever needed white-space trimming, but in principle this is a
backwards incompatible change. Let's just try this and see if anything
breaks.

(*1) The '\r' character was included in the .omr file, but as a raw
character. This patch now also escapes all characters < 32 as '&#xnn;'.