Preparations to support multiple openGL versions
We want to support 3 openGL versions:
 openGL ES 2.0: For Android and raspberry pi 3.
 openGL 2.1:    For older desktop hardware.
 openGL 3.3:    For integration with GTK (? not sure about the details)

The main reasons for this are:
- In principle openGL ES 2.0 offers all the features we need for
  openMSX, so only supporting this should be sufficient. However getting
  openGL ES to work on a windows desktop system seems non-trivial. And
  since we already have core openGL working on windows, using openGL 2.1
  (instead of openGL ES 2.0) might be an easier solution (workaround).
- During the openMSX-16.0 development cycle D15C0DE submitted some
  patches to upgrade to openGL 3.3 (instead of 2.1). OpenGL 3.3 was
  released in March 2010, so upgrading seemed reasonable. However after
  the openMSX-0.16 release we got reports that openMSX stopped working
  for some people. Because openMSX doesn't really require any of the new
  openGL features and because the speed gains are modest I decided to
  revert (part of) D15C0DE's work (and that resolved the reported
  problems).

I did not realize that the reason for D15C0DE's patches were to
integrate openMSX into some GTK based GUI (again not sure about the
details). For now(?) this work is closed source. Supporting out-of-tree
closed source projects is not a high priority, but:
- The extra code to support openGL 3.3 seems minimal (though it's not
  100% finished yet, see below).
- In the future we may want to upgrade to openGL ES 3.0, and the openGL
  3.3 code is a step in that direction.

This patch makes 3 main changes:
- The texture format 'GL_LUMINANCE' is no longer supported in openGL
  3.3. The workaround is to use the 'GL_RED' format and then setup
  appropriate GL_TEXTURE_SWIZZLE_RGBA texture parameters. This
  workaround is guarded by an #ifdef.
- Vertex Arrays (VAO) are not yet supported in openGL (ES) 2.x, but they
  seem to be required in openGL 3.3. Luckily it's sufficient to setup
  one global VAO and then ignore it. Of course this won't give any of
  the possible speedups of VAO, but in openMSX those speedups will
  anyway be small. This workaround is guarded by an #ifdef.
- While I was reverting D15C0DE's patches that introduce VAO, I
  overlooked that some of those patches also introduce more buffer
  objects (GL_ARRAY_BUFFER and GL_ELEMENT_ARRAY_BUFFER). Those are
  supported in openGL ES 2.0. So this patch re-adds those.

Currently selecting between the 3 supported versions requires a one line
code change (in src/video/GLUtil.hh). The default is still openGL 2.1.
But if I manually change this to openGL 3.3, then all rendering seems to
work again (and without triggering any openGL errors). This was not the
case before this patch.

This patch only changes the C++ code, it does not yet change any shader
code. Here's the commit message from D15C0DE's patch:
    Upgrade shaders to GLSL 3.30. Although the GLSL 1.10 shaders work on
    my system, technically their use is deprecated in OpenGL 3.3, so
    it's better to upgrade them.
So indeed we should also find a way to adjust the shader code. Maybe
that's possible with some #define's in the shader code? But in any case
that's lower priority and definitely something for a separate patch.