Use SDL_SetWindowFullscreen() to toggle fullscreen ..
.. instead of destroying/re-creating the whole window.

In revision 1e8f49e96a (Nov 2005) we explicitly went for this re-create
approach. Here's a comment explaining the reasoning:
   // In win32, toggling full screen requires opening a new SDL screen
   // in Linux calling the SDL_WM_ToggleFullScreen usually works fine
   // We now always create a new screen to make the code on both OSes
   // more similar (we had a windows-only bug because of this difference)

Though that was in 2005 with SDL1. Since then we switched to SDL2 and we
also switched to desktop-fullscreen-mode rather than 'true' fullscreen
mode (so upscaling instead of an actual video mode switch). Let's see if
using SDL_SetWindowFullscreen() (the equivalent of the old SDL1
SDL_WM_ToggleFullScreen() function) now works properly (it does work
fine for me in Linux).

The main motivation for this change is this bug:
    When switching to fullscreen on Pioneer PX-7, the videosource is changed
    https://github.com/openMSX/openMSX/issues/1254

See the bug report for the details, short summary: in the current code
re-creating the window implies re-creating all the VideoLayer objects
and that (temporarily) changes the allowed values for the 'videosource'
setting. And after this whole process that setting may no longer have
the same value as before the fullscreen switch. After this patch we keep
the same window and the same VideoLayer objects, and so we avoid this
videosource problem.

This is an experimental patch, needs more testing, especially on
windows.