Fix navigating to file pool for disks
When introducing drag-and-drop of folders, I broke this functionality. See
624996b. I was using the property 'is absolute path' to find out whether it's a
dropped folder (should be inserted as disk) or a folder to navigate to. But
that doesn't work for the file pool entry. These are also given as absolute
paths. Solution: explicitly mark file pool entries with a specific prefix and
check for that. Renamed abspath to dir_as_disk to make the code easier to
understand.