Fix filename-tab-completion
Was broken after revision cab2f971bc. The reason was that the new
foreach_file_xxx() utility functions always add a '/' character between
the given start-directory and the files/sub-directories in that
directory. Also when the start-directory already ended in a '/'
character.

Normally that's not a problem, except for tab-completion because it
relies on exact string matches.

Now we only add '/' when it doesn't already end in '/'. This was also
the behavior of the old 'ReadDir' utility.