Improve convenience for I/O only extensions
Allow I/O only extensions to be removed from cartridge slot menu for
convenience: I/O only slots can be viewed to be able to remove extensions from
them.

Also, make clear that an extension is added to an I/O slot or a "real" slot in
the message. This way the user knows up front that the cartridge slot he
selected will be unmodified.

(There is no way to let an I/O cart occupy a physical slot in openMSX. And it
would make things less convenient if such cartridges would occupy a slot.)