Show display names and make a bit more room for them
I.e. for what's inside the cartridge slot.

At the same time, simplify get_slot_content a bit, as we don't really have to
care about the different interface between cassetteplayer/laserdisc and other
media slots/commands, as the 2nd item in the output is either empty or
non-empty, indicating an empty or non-empty media slot.