Use c++20 [[no_unique_address]]
Normally in c/c++ each different variable needs to have a different
address.

Even in cases like this:
    struct IsOdd {
       bool operator()(int x) const { return x & 1; }
    };
    struct S {
        int x;
        IsOdd p;
        int y;
    };
The struct 'IsOdd' is an empty struct. That is: it contains no data. But
if it is used inside a larger struct 'S', because the variables 'x', 'p'
and 'y' must all have different addresses (and proper alignment),
sizeof(S) == 12, even though 'S' only needs 8 bytes of effective storage.

C++20 introduced the attribute [[no_unique_address]], that relaxes the
rule that each variable must have a different (=unique) address. So if
we change the code to:
    struct S {
        int x;
        [[no_unique_address]] IsOdd p;
        int y;
    };
then sizeof(S) is indeed only 8 bytes.

Zero-sized structs may seem weird, but in c++ they are fairly common.
Typical examples are functors (like the above 'IsOdd' predicate), or
(stateless) lambda functions.

Prior to c++20 there were workarounds to obtain the same storage
efficiency. But they are more complex, so in openMSX we only bothered to
use them in a few places.

This patch simplifies those workarounds and also applies
[[no_unique_address]] in some other places were it makes sense.