Fix bug in file browsing in OSD menu
There was confusion about absolute paths. If an item is a subdir of the current
working directory, it would trigger as an absolute path and insert it as
dir-as-disk. On Mac and Linux I could reproduce this by browsing to .openMSX
and then to share.
I never realized there was a Tcl proc to query the path type. With that I could
fix this. Still needs to be confirmed on Windows.