Do not filter identical names in plugged extension list
This doesn't make sense, if there are multiple with the same name, they are
actually there. It's not some effect of symlinks.

This is easily triggered by inserting the same extension multiple times. This
gives all kinds of other problems in this script, but that will be fixed in a
later commit.