Added info topic about extension instances
machine_info extension <extension> will now give information about the
extension instance: 1) the config file name (for XML based extensions) or the
ROM file name (for bare-ROM generated extensions) and the devices that are part
of the extension.
There was no other way yet to retrieve this information. Especially important
if the extension name is not the same as the file name, which happens when
inserting the same extension twice. E.g. inserting two times the 'fmpac'
extension gives both an 'fmpac' and an 'fmpac (1)' extension instance. From the
latter it was not robustly possible to find out which extension file was
actually used to create it.