Fixes for when extension name is not extension filename
The previous code wrongly assumed that the extension name is the same as the
file name of the extension, and thus that it could use
utils::get_extension_display_name_by_config_name to get a pretty printed
display name for an extension. But that assumption is wrong, see comment in
previous commit.

Instead, we need to use the new machine_info extension command to find out the
filename and use that with utils::get_extension_display_name_by_config_name.
This has now been put behind a proc get_extension_display_name.

With the same new command we can also easily get the ROM device name from the
bare ROM extension, so get_ROM_device_name_by_filename became obsolete.

When inserting an extension, we must catch the output of the ext command to
know which name the extension instance got. So make use of that to display the
pretty name on the UI.

As a bonus, display the slot behind the list of inserted extensions in the
Extension -> Remove menu.

And lastly, the duplicated way to get a pretty printed slot location of an
extension has been put behind a proc get_extension_slot_description.