Fix filename-tab-completion on empty string
In the implementation of foreach_dirent() we add a '/' character between
the given path and the directory entries in that path. We already fixed
that to only add that '/' when the path didn't already end on a '/'. Now
we fix that we don't need to add '/' when that path is an empty string
(because otherwise the result is a absolute path rather than an entry in
the current directory).

This sort-of happened to work on Linux because of an optimization where
we avoid calling getStat() in many cases. Windows doesn't implement that
optimization, so there it consistently failed.