Remove special bit-patterns from RomType enum
Before this patch we relied on specific numerical values for some of the
RomType enums. More in detail: for some types some of the bits encoded
some extra information.

This patch removes that. Now all RomType enums have consecutive values.
That allows to use this enum as an index in an array. This will be useful
for the next patch.