Create desktop and metainfo in Contrib/linux and update install.py
I did not add the changes for catapult, since I guess that should
go to the catapult package.

install.py should now do the right thing when 'make install' is
called, and install the icons, .desktop and .metainfo.xml files
in the right location. XDG_DATA_HOME check is mainly for flatpak
since it does not use /usr but /app, and this way it installs the
files in the correct location.