Use different numerical value for an invalid time_t
Before we used '-1', but apparently this does occur in practice as
modification date on files (and it triggers assertions in FilePoolCore).
(Possibly this wrong timestamp originates from rounding down a date
close to unix epoch??)

Now we use the most negative value. And we hide this decision in the
constant INVALID_TIME_T.

In the very unlikely case that there is a file with this timestamp, we
return 'INVALID_TIME_T + 1' for the modification date of that file.