Fixed potential array-out-of-range read in YM2413Okazaki
This is probably an old issue. Often but not always harmless (might
occasionally cause a click in the generated sound). But triggers an
assertion now in debug build because we switched from plain C-arrays to
std::array (for constexpr related stuff) and the latter does
bound-checks in debug build.

YM2413Okazaki is no longer used by default (replaced by NukeYKT), but
it's still used when replaying (older) .omr files.