Use (compile-time) perfect-hash function to lookup RomType names.
Translating rom-name to rom-type-enum was done via a hash-map, using a
'regular' hash-function. Good hash functions try to minimize the chance
that two different keys map to the same hash-value. When all possible
keys are known in advance (as is the case for rom-type-names (and
aliases)) then it's possible to construct a 'perfect' hash function.

A perfect hash function has the property that all the (valid) keys are
guaranteed to produce a different hash value. And with this guarantee
the structure of the hash-map can become simpler, because there's no
need anymore to resolve collisions.

Moreover all keys are already known at compile-time, so that allows to
construct the perfect-hash-function already at-compile time (before this
patch we initialized our hash-map at run-time on first use).

(In general) a perfect hash function does not come for free. While a
regular hash-function is often stateless, the perfect hash function we
construct here does require extra storage. Though only about 12-bits per
key in this specific case.

For more details see the comments in utils/MinimalPerfectHash.hh.