Applied clang-tidy modernize-use-nodiscard
Functions marked with [[nodiscard]] produce a compiler warning when they
get called, but their return value is ignored.

This clang-tidy analysis pass (conservatively) suggests that all
pure-functions (functions without side effects) should be marked
nodiscard.

But there are _many_ more functions where it makes sense to add this
annotation. I annotate some, but far from all.

For example: it's OK to ignore the result of
    byte MSXDevice::readMem(word address);
e.g. you may want to trigger the read for it's side effects (e.g. a
device with memory mapped IO). But it's not OK to ignore the result of
    byte MSXDevice::peekMem(word address);
because we designed peekMem() to be side-effect-free. As programmers we
can reason about this distinction, but for a tool like clang-tidy this
is very hard. And that's why today clang-tidy is very conservative with
it's suggestions.