Fix compile error on clang
The code compiled fine with gcc and msvc, but not with clang.

See here for some background info:
  https://stackoverflow.com/questions/46114214/lambda-implicit-capture-fails-with-variable-declared-from-structured-binding

Technically clang is correct here: this code shouldn't compile when
interpreting the c++17 standard strictly. Although there's no good
reason to disallow this code. In fact c++20 will allow it, and the only
change required for this in the standard was to remove the wording that
disallows it. See:
  http://open-std.org/JTC1/SC22/WG21/docs/papers/2019/p1091r3.html

Luckily it wasn't to difficult to workaround this (temporary) problem.