Fix nodiscard compiler warning
Normally the return type of 'SoundDevice::updateBuffer()' should be
ignored. So we marked this function [[nodiscard]]. However for the
special case that zero output samples are required, we are allowed to
ignore the result. (Even for zero samples we call the function because
internally some sound device emulation is sub-sample accurate).

Rather than removing the [[nodiscard]] annotation, I preferred to keep
it and instead make it explicit in the code that we are intentionally
ignoring the result.

For some reason gcc did not give this warning, but clang did. Not sure
why, maybe because of virtual functions??