Fix #1299 Reverse is very slow when having a tape from a large WAV file inserted
On reverse we create a new machine instance, including new disks/tapes.
So for a large WAV image, that whole image gets read from disk (and
filtered) on each reverse action.

Though WAV images are read-only, so the resulting image on each reverse
action will be exactly the same. And this allows to cache that result.