Replace stringToInt() with stringTo<int>()
Replace the functions
  StringOp::stringToInt()
  StringOp::stringToUint()
  StringOp::stringToUint64()
with a single templatized
  StringOp::stringTo<T>()       e.g. stringTo<int>
function, where 'T' can be any integral type.

Internally this is implemented via 'std::from_chars()' (new in c++17)
instead of via the family of 'strtol()' libc functions.

One advantage is that the input is no longer required to be a
zero-terminated 'const char*', now it's a 'string_view'. This allows to
elimate a copy in some places.

This new function should run slightly faster, but also because it has
the following limitations compared to 'strtol()':
- Leading whitespace is not accepted. I don't believe this is a problem
  for openMSX (e.g. when parsing numbers in XML we already strip
  whitespace at a higher level).
- Leading '+' sign is not accepted (not a problem).
- Octal notation (start with '0') is not recognized, instead it's
  interpreted as decimal. Not a problem, probably even a good thing.
- This function does NOT depend on the locale (not a limitation, but a
  good thing ;).

There's also a variant that doesn't do dynamic base detection. E.g.
  StringOp::stringToBase<10, unsigned>()
More on this in the next patch.