Fix crash when using .cas images
Broke after 703bbf3fc747 introduced a cache for WavImages.

When inserting a cassette image, we first try to parse it as a .wav
file, if that fails we try as .cas. Though the cache implementation in
WavImage had a bug: when parsing as .wav failed, it still created an
(invalid) entry in the cache. And so when using .cas images, we trigger
this bug.