Fix(?) HIGHDPI stuff on macos
On my system (Linux / Intel gfx) there seems to be a bug in
SDL_GL_GetDrawableSize() when switching from fullscreen back to desktop
mode. We (already) have a workaround for that in the code.

Though this workaround breaks HIGHDPI renderering. (There we do need to
query the size via SDL_GL_GetDrawableSize(), because it might be
different from the window-size). On Linux/windows HIGHDPI rendering
doesn't seem to have any effect(?) so, for now, only enable HIGHDPI on
macos, and only apply the workaround on non-macos.