Various small tweaks
For example change:
   for (int i = 0; i < 9; ++i) {
       buf1[i] = 0;
       buf2[i] = 0;
       buf3[i] = false;
   }
into:
   ranges::fill(buf1, 0);
   ranges::fill(buf2, 0);
   ranges::fill(buf3, false);

Whether it's better to have one loop which fills all arrays, or one loop
per array, depends on various things (size of the arrays, details of the
CPU architecture, ...). In any case better write the simplest code and
let the compiler decide to have one or multiple loops. When we use the
'range::fill()' algorithm we don't have to worry about the size of the
array (and all arrays having the same size).