Remove left/right border drawing optimization
From one frame to the next, the left/right border pixels very often look
the same. The following patch series took advantage of that:
    commit f340117a62ed073906b512765d6af22f07dfa31b
    Author: Wouter Vermaelen <vermaelen.wouter@gmail.com>
    Date:   Fri Jul 25 12:09:28 2014 +0200
        [1/2] Optimize left/right border drawing

Though in some circumstances we wrongly tried to reuse the left/right
border pixels to draw a full (bottom) vertical border line. This caused
  bug#1269: Flickering line gone after switching renderers

We can fix this by making the code aware of the difference between
left/right and top/bottom borders. Though I didn't do that, instead I
disabled the optimization.

The original patch series talked about a gain of 1%-2%. I have no reason
to doubt those numbers, but if I today re-measure the performance gain
caused by skipping drawing the left/right border it's only about
0.2%-0.4%.

I'm not exactly sure why there's such a large difference between the old
and the new measurement. A lot of openMSX code has changed, so maybe the
rendering code has become (relatively) less important? But likely a large
part of the explanation is that the old and new measurements are done on
different machines (e.g. the CPU in my new machine has much larger
caches).

In any case, based on the new measurements, the optimization doesn't
seem worth it anymore. Not enough gain for the extra complexity it
introduces. And it's also not a disaster if on some (older) machines we
now loose 1%-2%.