Use xrange() for simple counting for-loops
This code refactoring is a bit experimental. I think it's a good idea,
but it's hard to know for sure without actually trying it on a
non-trivial codebase and gaining some more experience with it. This
patch starts that experiment, and who knows, maybe in half a year I'll
have a different opinion.

We have many simple for-loops in the code of the form:
    for (int i = 0; i < NUM_LEDS; ++i) { ...
    for (unsigned i = 0; i < v.size(); ++i) { ...

This patch rewrites those to:
    for (auto i : xrange(NUM_LEDS)) { ...
    for (auto i : xrange(v.size())) { ...

The meaning of both versions is the same, and often the compiler even
generates the exact same (x86) code for them. Then why move away for
this typical c-style loop?

The c-style loop has some problems:
- In the first example the type of the constant 'NUM_LEDS' is not
  immediately visible. It's defined in a completely different part of
  the source code. It might be 'int', 'unsigned', 'size_t', ... Ideally
  the induction variable 'i' in the loop should have the same type. In
  the example I picked 'int', but from looking only at the loop you
  cannot tell if that's correct. In the xrange-variant we could declare
  the type as 'auto' and then the compiler will automatically deduce the
  correct type. (We can write 'auto i = 0' in the first variant, but
  that doesn't help: this deduces the type of '0' (which is 'int') not
  the type of 'NUM_LEDS').
- In the second example it's even more important to pick the correct
  type. If we assume 'v' is some container (e.g. a std::vector), then
  probably 'unsigned' is the wrong type, it likely should be 'size_t'.
  You could rightfully argue that this doesn't matter in practice: as
  long as there are no more than 4 billion element in the vector, using
  a 32bit rather than a 64bit induction variable should work fine.
  However this also forces the compiler to generate slower code because
  it explicitly requires the (wrong) wrapping behavior at the 32bit
  boundary. This either requires the compiler to emit extra 32->64 bit
  zero-extension instructions or prevents the compiler from promoting
  the integer-based induction variable to pointer-based.
- Another problem in the second example is that, in principle, the
  sub-expression 'v.size()' needs to be re-evaluated in every iteration.
  In some cases the compiler can prove that this is not needed, but not
  in general.
- The variants that use xrange() are often also shorter. By itself
  that's not a good reason to switch to it, but it's a bonus.

Compared to the traditional c-style for loop, the xrange for-loop has
some limitations:
- It can only do simple counting loops where the number of iterations is
  known at the start of the loop (exception: you can still have 'break'
  statements in the body of the loop).
So the xrange loop is strictly less powerful than the traditional loop.
But that's a good thing. It means we can use xrange for the simple loops
and traditional loops for the more tricky cases. And this makes the
tricky cases stand out more.