Add/use ranges::iota()
The previous patch refactored many (simple) for loops. Where possible it
also replaced some loops with an algorithm. For example:

Before:
    for (unsigned i = 0; i < 8; ++i) {
        blockNr[i] = 255;
    }
After:
    ranges::fill(blockNr, 255);

The latter is more descriptive. But it's also more robust: if later the
size of the blockNr[] array is changed we can't forget to also change
the interation-count in the loop. (The definition of the array, and the
loop are typically in two different source files).

We also had several loop of the form:
    for (int i = 0; i < 4; ++i) {
        memMapRegs[i] = i;
    }

C++20 added the algorithm 'std::iota' that does exactly this: fill a
range with an increasing value. (Languages like 'rust' also offer this
algorithm with the same name). I added it to our future-c++ emulation
layer. So now this example simplifies to:
    ranges::iota(memMapRegs, 0);

And this offers the same robustness advantages as the 'fill' example
above.