Add/use enumerate() to simplify some loops
Many programming languages offer a utility like enumerate(). It's not
too difficult to implement it in C++. This allows to simplify some
loops. For example:

Before:
    for (unsigned i = 0; i < MAX_DEVICES; ++i) {
        if (!devices[i].fs) return i;
    }
After:
    for (auto [i, dev] : enumerate(devices)) {
        if (!dev.fs) return i;
    }

Like in the previous patches this makes the code more robust against
changes in the array dimension. In this specific example it's solved by
the 'MAX_DEVICES' constant. By using enumerate that workaround is no
longer needed.