gl_{vec,mat} remove manual SSE optimizations
The type 'vecN<4, float>' (aka 'vec4') is mostly used in openGL related
calculations. For us the speed of these calculations is not super
important, but of course faster is always better. For this reason we
made sure that (on x86) the compiler emits SSE instructions to roughly
speedup these operations by a factor 4x. In order to accomplish this we
manually vectorized (parts of) the code using intrinsics.

Part of the implementation used this union:
    union {
        __m128 e;
        float e_[4];
    };
And sometimes we wrote data via one member of the union, but then read
it back via the other member. In practice this happened to work fine on
all compilers we use, but strictly speaking it's undefined behavior
according to the C/C++ standard.

I could fix this, but unfortunately I had to give up the following
syntax:
    vec4 v = ...
    // This no longer works:
       v[1] = 2.0;    // modify a mutable reference
    // This replacement works, but it's not as nice
       v.set(1, 2.0); // setter method
More technically, I couldn't find a (standard compliant) way to return
references to individual elements. Instead I had to use setter methods.

Taking a step back: is manually vectorizing this code still needed? This
code was originally written in May 2014 (that's around the time gcc-4.9
was released, now we have gcc-10). At that time these optimization
certainly helped. But if I test again today, I found that gcc/clang are
able to auto-vectorize this code pretty well. Not in all cases,
sometimes the manually optimized code is still better. But for the
functions that matter most (vector and matrix multiplication) both
versions are pretty similar. So there's no reason anymore to keep these
manual optimizations.