Make (most) gl{vec,mat,transform} functions constexpr
All these functions are pure mathematical functions (meaning they don't
have side effects). These are good candidates to mark 'constexpr'. This
allows the compiler to evaluate those functions in constexpr-context. So
roughly speaking: use these functions for compile-time calculations.

This only became possible after the previous patch which removes SSE
intrinsics.

At the moment we don't need this yet, but it opens possibilities for the
future.

One practical problem: some functions (vector normalization and rotation
matrices) use sqrt(), sin() and cos() internally. These are not (yet)
constexpr in standard c++. We do have constexpr re-implementations, but
those are a little more difficult to use, so for now I didn't bother.
But if it's needed in the future, we can do it.