Fix scriptablity of 'machine_info output_port'
When multiple devices share the same input/output ports you can query
the individual devices with a Tcl command like:

   <select turbor machine>
   <insert the 'fmpac' extension>
   lindex [machine_info output_port 0x7c] 0
--> returns  "MSX Music"

In other words the 'output_port' info topic returns a list of device
names.

However when there is only one device associated with a certain port and
you query that, the result is formatted as a string instead of as a Tcl
list. This makes a difference when the (single) name contains a space
like in the example above.

This patch fixes the code to always format the result as list, also when
that list only contains a single element.