Add proper rounding for double->EmuDuration conversion
This matters most for very short durations, e.g. the duration of one
Y8950 sample (1/49716 seconds).

This fixes
   bug#1308 Audio samples sound distorted.

That bug was a regression introduced in:
    commit 945f4ec
    Author: m9710797 vermaelen.wouter@gmail.com
    Date: Sat Aug 15 09:46:13 2020 +0200
    Tweak calculation of ResampledSoundDevice::emuClock speed

That patch switched from using DynamicClock::setFreq() to
DynamicClock::setPeriod() because the latter introduces less rounding
error (which is true). But ironically calculating that period-parameter
was done imprecise. That calculation involved a conversion from double
to EmuDuration. This patch adds proper rounding for that conversion.

Apparently the 'MC Hammer Music Demo' described in the bug report is
very sensitive to having the correct adpcm playback speed.