Lazily create 'Mixer' object
Before this patch we would:
- Create the 'Reactor'. This also (eagerly) creates the 'Mixer'. And in
  turn this creates the 'SDLSoundDriver' which opens the SDL sound
  stuff.
- Then we load the 'settings.xml' file.
- When some of the mixer-related settings (e.g. 'samples' or
  'frequency') differ from the default, the 'Mixer' picks up the new
  values. This creates a new 'SDLSoundDriver': the old one gets
  destroyed, this closes the SDL sound stuff, and the new one reopens
  it.

So during startup the sound stuff got initialized twice. This is
actually an old bug report:
    #784 [sf#390] [Bug] Audio device opened twice on startup
(This is not a real 'bug', but it is inefficient.)

This patch makes creating of the 'Mixer' object lazy. That is we wait
with creating it until someone actually requires it. In practice that
means it only gets created when the first 'MSXMotherBoard' object gets
created. At that time the 'settings.xml' file is already parsed. And
thus the sound stuff can immediately be initialized with the correct
settings.

In the past we initialized almost all Reactor sub-objects in a lazy way.
Though because these sub-objects have non-trivial inter-dependencies and
because there are many variations in the startup-sequence (e.g. start
normally vs start via '-control') this lazy initialization often was the
source of bugs (e.g. lead to circular dependencies). Because of that we
switched from (mostly) lazy to (mostly) eager initialization. This patch
takes a small step back. I verified that this is safe.