Remove 'commands' map from GlobalCommandController
'GlobalCommandController' kept a (hash)map from the command-name to the
corresponding 'Command' object. Though (apart from asserts) this map was
only used to implement the hasCommand() method. And that method was only
used in various commandline parsing functions: for example while parsing
'-cassetteplayer' to give an error when the selected machine has no
cassette port.

In other words the GlobalCommandController::hasCommand() function is not
critical. We can as well query the existence of a command via the Tcl
interpreter itself. So this patch removes that map from
GlobalCommandController. (But we keep it in debug builds, because the
extra checks it enables can be useful).

Note that MSXCommandController has a similar map, but that one is also
used to implement 'ProxyCommand' efficiently (e.g. to delegate from the
global proxy 'diska' command to the actual machine specific
'machine1::diska' command). So we keep that map.