Fix for multiple <filename> tags inside <rom>
Revision bcba71fbe added support for having multiple <filename> tags
inside a <rom> tag. Similar to how there can be multiple <sha1> tags.
None of our machine/extension configs use this, but apparently it is
used internally for the VLM5030 rom to load the voice rom from two
possible locations:

    romElement.addChild( // load by predefined filename in software rom's dir
        "filename", strCat(FileOperations::stripExtension(romFilename), "_voice.rom"));
    romElement.addChild( // or hardcoded filename in ditto dir
        "filename", "keyboardmaster/voice.rom");

Unfortunately the code had a bug: all filenames would be tried (in
order), but the results of all but the last one are rejected. So this
means that it still only worked properly for either no or exactly one
<filename> tag.

The fix is to stop at the first successfully loaded <filename>.