Tweak FileOperations::getNativePath()
On windows the path delimiter is typically a back-slash while on other
systems it's typically a forward-slash. For internal manipulation in the
code it's easier to work with a single convention. Therefor we have two
helper functions in FileOperations: getConventionalPath() and
getNativePath(). On windows these will convert back- to forward-slashes
and the other way around. On other systems these don't do anything.

Actually functionally these 'don't do anything', but they still make a
copy of the given string. In some cases that copy can be avoided. So
that's what this patch does.

For some reason we already applied this tweak to getConventionalPath()
but not yet to getNativePath(), this patch fixes that inconsistency.