Optimize usage of FileOperation::expandTilde()
Tilde-expansion is an idempotent operation. That is doing it just once,
or more than once has the same effect. Though on some code paths we did
perform this operation multiple times.

The idea is to (only) do tilde-expansion close to the place where
user-input is processed (e.g. command line parsing, or Tcl command
parsing), but no longer in any deeper code.

This patch takes some steps in that direction, but the work is not
finished yet. So I expect some follow up patches (and in fact this patch
is already a continuation of previous work).