Avoid string copy to store description string-literals
The following classes (and related/derived classes) take a 'description'
constructor parameter:
  Debuggable, Probbe, Setting, SoundDevice, TclCallback
That description was stored in a std::string object. That means it
stores a copy of the parameter, but in 99% of the cases that parameter
is a string-literal and then making a copy is not actually needed.

We could change these classes to store a string_view instead of string.
And that would get rid of the unnecessary copies. Unfortunately that
would lead to fragile code, because e.g. passing a temporary std::string
would still compile fine, but would create a dangling references to the
(soon to be deleted) temporary string.

This patch introduces the 'static_string_view' class. It acts mostly
like a regular string_view, except that it can only be constructed from
a string-literal. String-literals have static lifetime (meaning they
live for the whole program duration). And thus they are free of this
dangling-reference problem. (Not 100% true, but close enough. See
unittest for more details).

The first paragraph talked about 99% of the arguments being
string-literals. The last 1% could either be changed to be a
string-literal. Or it was a user-defined-setting: these settings
(including their description) are defined via Tcl scripts and therefor
it cannot be c++ string-literals. This patch also adds some
infrastructure to both refer to the string via a reference and safely
manage the lifetime/storage of the string.

This patch eliminates several hundreds (depending on the exact MSX model)
string copies during openMSX startup.