In sha1 code, replace ptr+size with span<>
A pointer plus an integer parameter could mean lots of thing. By
combining them into a single span<> parameter it's semantically clear
that you're passing (a view on) a memory buffer.

In some cases this allows to simplify the caller code. (In the future)
this extra semantic information can allow static analyzers to do a
better job.