Optimize SHA1::finalize()
Two main tweaks:

Instead of adding (up to 63) individual zero-bytes via calls to update()
in a loop, directly memset() the (1 or 2) parts of the internal buffer
to zero.

Instead of shifting/masking 8 individual bytes of a 64 bit value (and
store via 8 update() calls), convert the value to big-endian order and
write (memcpy()) it directly to the internal memory buffer.

Extended the unittest to cover all corner cases.