Added Setting::getOptionalValue() ...
... as an alternative for Setting::getValue(). In case of an error the
former returns an empty std::optional, the latter throws an exception.

During construction of ProxySettings, it's quite common for getValue()
to report an error. And the scope in which the getValue() gets called is
the same as the scope that handles the error. So this case it's more
convenient to use std::optional instead of c++ exceptions. (And we also
avoid constructing the exception message string which will anyway be
ignored).