Tweak parsing of 'after' command
The after command has several forms:
  after time <seconds> <command>
  after frame <command>
but also
  after <n> <command>          (for backwards compatibility, similar to 'after time')
  after <event> <command>

During parsing (of these last two) we first try to parse the 2nd
argument as an integer then we try to parse as an event. Most likely
parsing as integer will fail.

Before this patch trying to parse as integer could only be done via c++
exceptions: execute in a try-block and catch the exception. But when the
error-case is expected rather than exceptional that's not the best
approach. In the first place because of readability, but also for
performance.

This patch adds:
    std::optional<int> TclObject::getOptionalInt() const;
as an alternative for:
    int TclObject::getInt(Interpreter& interpreter) const; // possibly throws

And that allows to simplify/optimize:
    try {
        int time = tokens[1].getInt(getInterpreter());
        handle(time);
    } catch (CommandException& /*e*/) { // ignore message
        try {
            ...
        } catch (...) {
            ...
        }
    }
to:
    if (auto time = tokens[1].getOptionalInt()) {
        handle(*time);
    } else if (...) {
      ...
    } else {
       ...
    }